+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 ----
 * use ``concurrent.futures`` instead of ``multiprocessing``
 * make some dependencies optional (Flask, boto3, etc.)
+* speed up ``count_tiles()``
 
 
 ----